sensorTransmissions-service/requirements.txt
flask
psycopg2-binary
sqlalchemy
pika